Calculate CCFs

Calculate CCFs

Calculate CCFs

Functions

Types

Internals

Calculate CCFs

Functions

EchelleCCFs.calc_ccf_chunk — Method

calc_ccf_chunk( chunk, ccf_plan ) Convenience function to compute CCF for one chunk of spectrum.

Inputs:

chunk

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and line list from ccf plan

EchelleCCFs.calc_ccf_chunklist — Method

calc_ccf_chunklist ( chunklist, var_list, ccf_plans ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

var_list: vector of variance vectors for each chunk

ccf_palns: vector of ccf plans (one for each chunk)

Optional Arguments:

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the line list and mask_shape from the ccf plan for each chunk.

EchelleCCFs.calc_ccf_chunklist — Method

calc_ccf_chunklist ( chunklist, ccf_plans ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

vector of ccf plans (one for each chunk)

Optional Arguments:

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the line list and mask_shape from the ccf plan for each chunk.

EchelleCCFs.calc_ccf_chunklist_timeseries — Method

calc_ccf_chunklist_timeseries( chunklist_timeseries, line_list ) Convenience function to compute CCF for a timeseries of spectra, each with a chunklist. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

Optional Arguments:

ccf_plan (BasicCCFPlan())

verbose (false)

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the ccfplan. Note that the ccfplan provided is used as a template for creating a custom ccfplan for each chunk that only includes lines that reliably appear in that order for all spectra in the chunklisttimeseries.

EchelleCCFs.calc_ccf_v_grid — Method

calc_ccf_v_grid( plan ) Return range with 2n+1 points between -vmax and vmax where CCF is to be evaluated. Units based on those in plan.

EchelleCCFs.calc_length_ccf_v_grid — Method

calc_length_ccf_v_grid( plan ) Return number of points in the velocity grid (without needing to create the range).

EchelleCCFs.ccf_1D! — Method

`ccf_1D!(ccf_out, λs, fluxes; ccf_plan )`

Compute the cross correlation function of a spectrum with a mask. Generalized version that should work with different mask shapes.

Inputs:

ccf_out: 1-d array of size length(calcccfv_grid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

plan: parameters for computing ccf (BasicCCFPlan())

EchelleCCFs.ccf_1D! — Method

`ccf_1D!(ccf_out, ccf_var_out, λs, fluxes, var; ccf_plan )`

Compute the cross correlation function of a spectrum with a mask. Generalized version that should work with different mask shapes.

Inputs:

ccf_out: 1-d array of size length(calcccfv_grid(plan)) to store output

ccf_var_out: 1-d array of size length(calcccfv_grid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

var: 1-d array of flux variances

Optional Arguments:

plan: parameters for computing ccf (BasicCCFPlan())

EchelleCCFs.ccf_1D — Method

`ccf_1D( λs, fluxes; ccf_plan )`
Compute the cross correlation function of a spectrum with a mask.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

ccf_plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

EchelleCCFs.ccf_1D — Method

`ccf_1D( λs, fluxes, vars; ccf_plan )`

Compute the cross correlation function of a spectrum with a mask and its variance. This version uses pixel variances. WARNING: Still experimenting. Needs testing.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

ccf_plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

Types

EchelleCCFs.AbstractCCFMaskShape — Type

A struct implementing a specific mask shapes should be a subtype of AbstractCCFMaskShape.

EchelleCCFs.AbstractCCFPlan — Type

A struct implementing a specific plans describing where the CCF is to be evaluated should be a subtype of AbstractCCFPlan.

EchelleCCFs.AbstractLineList — Type

A struct implementing a line list should be a subtype of AbstractLineList.

EchelleCCFs.BasicCCFPlan — Type

Basic plan for computing the CCF roughly between vcenter-vmax and vcenter+vmax with step size v_step.

EchelleCCFs.BasicCCFPlan — Method

BasicCCFPlan

Optional arguments:

midpoint: (default_v_center)

step: (default_v_step)

max: (default_v_max)

EchelleCCFs.BasicLineList — Type

A basic line list for passing to compute CCFs. Contains (views into) arrays specifying the minimum and maximum wavelength range and weight for each line.

EchelleCCFs.BasicLineList — Method

BasicLineList( λ, weight )

EchelleCCFs.CosCCFMask — Type

CosCCFMask Cosine mask with one parameter, it's quarter period, i.e., where to truncate it, as a velocity in m/s. Mask weights are stored separately in a line list.

EchelleCCFs.CosCCFMask — Method

Functor for returning PSF for Δv <= half_width.

EchelleCCFs.CosCCFMask — Method

CosCCFMask( inst )

EchelleCCFs.GaussianCCFMask — Type

GaussianCCFMask A truncated Gaussian mask with two parameters, it's stdandard deviation and where to truncate it, both as a velocity in m/s. Mask weights are stored separately in a line list.

EchelleCCFs.GaussianCCFMask — Method

Functor for returning PSF for Δv <= half_width.

EchelleCCFs.GaussianCCFMask — Method

GaussianCCFMask( inst ; scale_factor )

EchelleCCFs.SuperGaussianCCFMask — Type

SuperGaussianCCFMask A truncated Gaussian mask with two parameters, it's stdandard deviation and where to truncate it, both as a velocity in m/s. Mask weights are stored separately in a line list.

TODO: Repalce Gaussian with super-Gaussian Warning: Not implemented/tested yet.

EchelleCCFs.SuperGaussianCCFMask — Method

SuperGaussianCCFMask( inst ; scale_factor )

EchelleCCFs.SuperGaussianCCFMask — Method

Functor for returning PSF for Δv <= half_width.

EchelleCCFs.TopHatCCFMask — Type

TopHatCCFMask The standard tophat mask with one parameter, it's half width as a velocity in m/s. Mask weights are stored separately in a line list.

EchelleCCFs.TopHatCCFMask — Type

TopHatCCFMask( ; half_width=default_v_width )`

EchelleCCFs.TopHatCCFMask — Method

Functor for returning 1 for any Δv <= width.

EchelleCCFs.TopHatCCFMask — Method

TopHatCCFMask( inst )

Internals

Base.length — Method

Return length of line linst.

EchelleCCFs.calc_ccf_and_var_chunk — Method

calc_ccf_and_var_chunk( chunk, ccf_plan ) Convenience function to compute CCF for one chunk of spectrum.

Inputs:

chunk

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and line list from ccf plan

EchelleCCFs.calc_ccf_and_var_chunklist — Method

calc_ccf_and_var_chunklist ( chunklist, ccf_plans ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

vector of ccf plans (one for each chunk)

Optional Arguments:

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the line list and mask_shape from the ccf plan for each chunk.

EchelleCCFs.calc_doppler_factor — Method

calc_doppler_factor(vel)` Compute the longitudinal relativistic doppler factor given a velocity in meters per second.

EchelleCCFs.calc_order_ccf_chunklist_timeseries — Method

calc_order_ccf_chunklist_timeseries( chunklist_timeseries, ccf_plan ) Convenience function to compute separate CCFs for each chunk (typically an order) of each spectrum in a timeseries. CCF is evaluated using line list and mask_shape provided by the ccf plan for each chunk. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

A 3-d array containing the CCF at each (velocity, chunk, spectrum) Note that the ccfplan provided is used as a template for creating a custom ccfplan for each chunk that only includes lines that reliably appear in that order for all spectra in the chunklist_timeseries.

EchelleCCFs.calc_order_ccfs_chunklist — Method

calc_order_ccfs_chunklist ( chunklist_timeseries, line_list ) Convenience function to compute separate CCFs for each chunk (typically an order) in a spectrum. CCF is evaluated using line list and mask_shape provided by the ccf plan for each chunk.

Inputs:

chunklist_timeseries

vector of ccf plans (one for each chunk)

Return:

A 2-d array containing the CCF at each (velocity, chunk)

EchelleCCFs.ccf_1D_exper! — Method

`ccf_1D_exper!(ccf_out, ccf_var_out, λs, fluxes, vars; ccf_plan )`

Compute the cross correlation function of a spectrum with a mask. Generalized version that should work with different mask shapes. Weights each pixel by provided variance. WARNING: Still experimenting. Needs testing.

Inputs:

ccf_out: 1-d array of size length(calcccfv_grid(plan)) to store output

ccf_var_out: 1-d array of size length(calcccfv_grid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

var: 1-d array of flux variances

Optional Arguments:

plan: parameters for computing ccf (BasicCCFPlan())

EchelleCCFs.find_overlapping_chunks — Method

Return indices of any chunks in df that have overlapping lambda_hi[i] and lambda_lo[i+1].

EchelleCCFs.merge_chunks — Method

Return DataFrame with specifications for each chunk which will contain one or more lines. Input: line_list a DataFrame with: - lambda_lo, lambda_hi, lambda, depth Output: DataFrame with - lambda_lo & lambda_hi: boundaries for chunk - lambda & line_depths: arrays with info about each line

EchelleCCFs.project_mask! — Method

project_mask!( output, λs, ccf_plan; shift_factor )

Compute the projection of the mask onto the 1D array of wavelengths (λs) at a given shift factor (default: 1). The mask is computed from the ccfplan, including a linelist and maskshape (default: tophat). Assumes plan.linelist is already sorted.

EchelleCCFs.read_linelist_espresso — Method

Read line list in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

EchelleCCFs.read_linelist_vald — Method

Read line list in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

EchelleCCFs.read_mask_espresso — Method

Read mask in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

EchelleCCFs.read_mask_vald — Method

Read mask in VALD csv format. VALD format: lambda_lo, lambdaa_hi and depth.

EchelleCCFs.λ_air_to_vac — Method

Convert air wavelength (in Å) to vacuum wavelength Ref: https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion VALD3 tools use the following solution derived by N. Piskunov

EchelleCCFs.λ_vac_to_air — Method

Convert vacuum wavelength (in Å) to air wavelength Ref: Donald Morton (2000, ApJ. Suppl., 130, 403) via https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion